Drop PHP <7.1 and Symfony <5.3 support
