+++ .github/workflows/ci.yml
@@ -13,8 +13,8 @@ jobs:
             matrix:
                 include:
                     # Lowest Deps
-                    - php: 7.1
-                      symfony: 4.4.*
+                    - php: 7.2
+                      symfony: 5.4.*
                       composer-flags: '--prefer-stable --prefer-lowest'
                       can-fail: false
                     # LTS with latest stable PHP
@@ -24,17 +24,12 @@ jobs:
                       can-fail: false
                     # Stable Symfony branches
                     - php: 7.4
-                      symfony: 4.4.*
+                      symfony: 5.4.*
                       composer-flags: '--prefer-stable'
                       can-fail: false
-                    - php: 8.0
-                      symfony: 6.0.*
-                      composer-flags: '--prefer-stable'
-                      can-fail: false
-                    # Development Symfony branches
                     - php: 8.1
-                      symfony: 6.1.*@dev
-                      composer-flags: ''
+                      symfony: 6.1.*
+                      composer-flags: '--prefer-stable'
                       can-fail: false
 
         name: "PHP ${{ matrix.php }} - Symfony ${{ matrix.symfony }}${{ matrix.composer-flags != '' && format(' - Composer {0}', matrix.composer-flags) || '' }}"
